+++ src/airflow_local_debug/__init__.py
@@ -67,11 +67,12 @@ from airflow_local_debug.bootstrap import (
 )
 from airflow_local_debug.compat import get_airflow_version
 from airflow_local_debug.config_loader import get_default_config_path, load_local_config
+from airflow_local_debug.deferrables import detect_deferrable_tasks
 from airflow_local_debug.env_bootstrap import bootstrap_airflow_env
 from airflow_local_debug.graph import format_dag_graph, print_dag_graph, render_dag_svg, write_dag_svg
 from airflow_local_debug.live_trace import live_task_trace
 from airflow_local_debug.mocks import TaskMockRule, load_task_mock_rules, task_mock_rules_from_payload
-from airflow_local_debug.models import DagFileInfo, LocalConfig, RunResult, TaskMockInfo, TaskRunInfo
+from airflow_local_debug.models import DeferrableTaskInfo, DagFileInfo, LocalConfig, RunResult, TaskMockInfo, TaskRunInfo
 from airflow_local_debug.plugins import (
     AirflowDebugPlugin,
     ConsoleTracePlugin,
@@ -125,7 +126,9 @@ __all__ = [
     "debug_dag_cli",
     "debug_dag_file_cli",
     "debug_dag_from_file",
+    "DeferrableTaskInfo",
     "DebugPluginManager",
+    "detect_deferrable_tasks",
     "DagFileInfo",
     "DoctorCheck",
     "DoctorResult",